+++ examples/.lit_test_times.txt
@@ -1,0 +1,11 @@
+2.121925e-05 ast.toy
+1.192093e-06 basic.toy
+2.956390e-05 codegen.toy
+2.121925e-05 empty.toy
+2.408028e-05 jit.toy
+4.053116e-06 scalar.toy
+2.050400e-05 struct-ast.toy
+3.838539e-05 struct-codegen.toy
+3.170967e-05 transpose_transpose.toy
+3.576279e-05 trivial_reshape.toy
+2.479553e-05 user_defined_function.toy
